+++ src/components/UI/index.ts
@@ -2,3 +2,5 @@ export * from './Avatar/Avatar'
 export * from './Link/Link'
 export * from './Table/Table'
 export * from './Tabs/Tabs'
+export * from './ThemeToggle'
+
