+++ src/components/index.ts
@@ -1,7 +1,7 @@
-export * from './Header/Header'
-export * from './UserAlbums/UserAlbums'
-export * from './UserInfo/UserInfo'
-export * from './UserList/UserList'
-export * from './UserMedia/UserMedia'
-export * from './UserPosts/UserPosts'
-export * from './UserTabs/UserTabs'
+export * from './header/Header'
+export * from './userAlbums/UserAlbums'
+export * from './userInfo/UserInfo'
+export * from './userList/UserList'
+export * from './userMedia/UserMedia'
+export * from './userPosts/UserPosts'
+export * from './userTabs/UserTabs'
