+++ src/providers/ThemeProvider/index.tsx
@@ -1,4 +1,3 @@
-// contexts/ThemeContext.tsx
 import React, { createContext, useContext, useEffect, useState } from 'react'
 
 type Theme = 'light' | 'dark'
