+++ src/providers/index.ts
@@ -1,1 +1,2 @@
 export * from './RouterProvider'
+export * from './ThemeProvider'
